+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Dependency updates.
+
 ## [v9.0.3](https://github.com/lexicalunit/spellbot/releases/tag/v9.0.3) - 2023-08-03
 
 ### Changed